requirements.txt
bcrypt==5.0.0
cffi==2.0.0
cryptography==46.0.5
invoke==2.2.1
markdown-it-py==4.0.0
mdurl==0.1.2
netmiko==4.6.0
ntc_templates==9.0.0
paramiko==4.0.0
pycparser==3.0
Pygments==2.19.2
PyNaCl==1.6.2
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.0
pyserial==3.5
PyYAML==6.0.3
rich==14.3.3
ruamel.yaml==0.19.1
scp==0.15.0
textfsm==2.1.0
pywinrm==0.4.3
python-nmap==0.7.1
ansible==8.6.1
ansible-core==2.15.10
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
Markdown==3.10.2
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
pillow==12.1.0
pyparsing==3.3.2
pyspnego==0.12.1
python-dateutil==2.9.0.post0
qtwidgets==1.1
requests==2.32.5
requests_ntlm==1.3.0
resolvelib==1.0.1
scapy==2.6.1
six==1.17.0
sspilib==0.5.0
urllib3==2.6.3
xmltodict==1.0.4
